+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Add method to construct dictionary by removing a key.
   * Add method to filter entries from a dictionary.
   * Add alternative URL Query parameter encoding implementation.
+  * Add method to determine if one dictionary is a subset of another.
 
 2.4.0, 26-June-2014